+++ crates/ui-components/src/lib.rs
@@ -123,6 +123,7 @@ pub mod layout;
 pub mod legend;
 pub mod link;
 pub mod link_button;
+pub mod list;
 pub mod listbox;
 pub mod listbox_item;
 pub mod listbox_section;